+++ src/app/(providers)/(root)/(public)/join/_components/validationUtils.ts
@@ -1,0 +1,3 @@
+// 정규표현식 이메일과 비밀번호 유효성검사
+export const emailValidation = /^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,6}$/;
+export const passwordValidation = /^(?=.*[0-9])(?=.*[!@#$%^&*])[a-zA-Z0-9!@#$%^&*]{8,}$/;
